单位管理、会议座次、会议投票、会议纪要

+++ cqhc-system/src/main/java/com/cqhc/modules/meeting/domain/MeetingType.java
@@ -1,6 +1,5 @@
 package com.cqhc.modules.meeting.domain;
 
-import com.cqhc.modules.system.domain.Dept;
 import com.cqhc.modules.system.domain.Unit;
 import lombok.Data;
 import javax.persistence.*;
@@ -31,7 +30,7 @@ public class MeetingType implements Serializable {
     * */
     @NotNull
     @Column(name = "sort",nullable = false)
-    private Integer sort;
+    private short sort;
 
     /*
     * 名称
